requirements.txt
python-lorem
gmpy2
sympy
secrets
